tienda  proyecto

+++ Avance1Proyecto/js/clientes.js
@@ -4,13 +4,17 @@ import { supabase } from "./supabase.js";
 const tbody = document.getElementById("tbodyClientes");
 const btnLoad = document.getElementById("btnLoad");
 const btnAdd = document.getElementById("btnAdd");
+const btnCancel = document.getElementById("btnCancel");
 
 const txtId = document.getElementById("txtId");
 const txtNombre = document.getElementById("txtNombre");
 const txtCorreo = document.getElementById("txtCorreo");
 const txtTelefono = document.getElementById("txtTelefono");
+const tituloForm = document.getElementById("tituloForm");
 
 // FUNCIONES
+// VARIABLES GLOBALES
+let rolUsuario = "";
 
 // CONSULTAR
 const consultarClientes = async () => {
@@ -30,6 +34,17 @@ const consultarClientes = async () => {
 
   data.forEach((c) => {
 
+    const botonesAdmin = rolUsuario === "admin"
+      ? `
+        <button class="btnEditar btn btn-warning btn-sm" data-id="${c.id}">
+          Editar
+        </button>
+        <button class="btnEliminar btn btn-danger btn-sm" data-id="${c.id}">
+          Eliminar
+        </button>
+      `
+      : "";
+
     const tr = document.createElement("tr");
 
     tr.innerHTML = `
@@ -37,20 +52,12 @@ const consultarClientes = async () => {
       <td>${c.nombre}</td>
       <td>${c.correo}</td>
       <td>${c.telefono}</td>
-      <td>
-        <button class="btnEditar btn btn-warning btn-sm" data-id="${c.id}">
-          Editar
-        </button>
-        <button class="btnEliminar btn btn-danger btn-sm" data-id="${c.id}">
-          Eliminar
-        </button>
-      </td>
+      <td>${botonesAdmin}</td>
     `;
 
     tbody.appendChild(tr);
 
   });
-
 };
 
 // GUARDAR
@@ -103,7 +110,8 @@ const guardarCliente = async () => {
 
 // ELIMINAR
 const eliminarCliente = async (id) => {
-
+  if (rolUsuario !== "admin") return;
+ 
   if (!confirm("¿Eliminar cliente?")) return;
 
   const { error } = await supabase
@@ -126,17 +134,19 @@ const limpiarFormulario = () => {
   txtNombre.value = "";
   txtCorreo.value = "";
   txtTelefono.value = "";
+  btnAdd.textContent = "Guardar";
+  tituloForm.textContent = "Agregar Cliente";
 };
 
 // EVENTOS
 if (btnLoad) btnLoad.addEventListener("click", consultarClientes);
 if (btnAdd) btnAdd.addEventListener("click", guardarCliente);
-
+btnCancel.addEventListener("click", async () => limpiarFormulario());
 // DELEGACIÓN
 tbody.addEventListener("click", async (event) => {
-
+   
   const target = event.target;
-
+  if (rolUsuario !== "admin") return;
   if (target.classList.contains("btnEliminar")) {
     const id = target.getAttribute("data-id");
     await eliminarCliente(id);
@@ -162,11 +172,40 @@ tbody.addEventListener("click", async (event) => {
     txtNombre.value = data.nombre;
     txtCorreo.value = data.correo;
     txtTelefono.value = data.telefono;
+    btnAdd.textContent = "Actualizar";
+    tituloForm.textContent = "Editar Cliente";
   }
 
 });
 
 // INICIALIZACIÓN
-window.onload = () => {
-  consultarClientes();
+window.onload = async() => {
+   await obtenerRol();
+  console.log("Rol detectado:", rolUsuario);
+  await consultarClientes();
+};
+// OBTENER ROL USUARIO
+const obtenerRol = async () => {
+  const { data: userData } = await supabase.auth.getUser();
+
+  if (!userData?.user) {
+    window.location.href = "login.html";
+    return;
+  }
+
+  const { data, error } = await supabase
+    .from("usuarios")
+    .select("rol")
+    .eq("correo", userData.user.email)
+    .maybeSingle();
+
+  if (error) {
+    console.error(error);
+    return;
+  }
+
+  rolUsuario = (data?.rol || "vendedor").trim().toLowerCase();
+
+  console.log("EMAIL:", userData.user.email);
+  console.log("ROL BD:", data);
 };
